+++ benches/toks.rs
@@ -5,7 +5,7 @@ use {
     glassbench::*,
 };
 
-static PATTERNS: &[&str] = &["", "a", "br,o,c", "e,jenc,arec,ehro", "broot"];
+static PATTERNS: &[&str] = &["a", "réveil", "bro,c", "e,jenc,arec,ehro", "broot"];
 
 fn bench_score_of_toks(gb: &mut Bench) {
     for pattern in PATTERNS {